// frontend/src/lib/api/generated/zod.gen.ts
export const zProfile = z.object({
    id: z.string(),
    email: z.string(),
    display_name: z.string(),
    role: zRole,
    permissions: z.array(zPermission),
    sensor_ids: z.array(z.string())
});
// ...
export const zTokenPair = z.object({
    access_token: z.string(),
    token_type: z.literal('bearer'),
    access_expires_in: z.int(),
    refresh_token: z.string(),
    refresh_expires_in: z.int(),
    user: zProfile
});
// ...
export const zRefreshResponse = zTokenPair;